rename constant routes file
